requirements.txt
requests ~= 2.20.0
GroupyAPI == 0.10.3